requirements.txt
Django>=3.0
psycopg2>=2.8
djangorestframework>=3.11.2
django-apscheduler>=0.6.2
APScheduler>=3.10.4
requests>=2.28.1
gunicorn>=20.1.0